requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic
psycopg2-binary
